Refactor ModelLoader to use PrefillInputsMixin for input handling
- Updated the ModelLoader class across multiple implementations (Gemma, Llama, Qwen 2.5) to inherit from PrefillInputsMixin, centralizing the logic for loading prefill-step inputs.
- Removed redundant load_inputs_prefill methods from individual ModelLoader classes, enhancing code maintainability and reducing duplication.
- Updated imports to include the new PrefillInputsMixin, ensuring compatibility with the refactored input handling approach.
