+++ qwen_2_5/causal_lm/pytorch/loader.py
@@ -20,8 +20,8 @@ from ....config import (
     Framework,
     StrEnum,
 )
-from ....tools.prefill_inputs import get_prefill_texts_for_batch, PREFILL_TEXTS
-from ....tools.utils import cast_input_to_type, get_static_cache_decode_inputs
+from ....tools.utils import get_static_cache_decode_inputs
+from ....tools.prefill_inputs import PrefillInputsMixin
 
 
 class ModelVariant(StrEnum):
@@ -45,7 +45,7 @@ class ModelVariant(StrEnum):
     QWEN_2_5_MATH_7B = "Math_7B"
 
 
-class ModelLoader(ForgeModel):
+class ModelLoader(ForgeModel, PrefillInputsMixin):
     """Qwen 2.5 model loader implementation for causal language modeling tasks."""
 
     # Dictionary of available model variants using structured configs
@@ -315,33 +315,6 @@ class ModelLoader(ForgeModel):
         )
 
         return self.config
-
-    def load_inputs_prefill(self, dtype_override=None, batch_size=1, seq_len=128):
-        """Load prefill-step inputs with texts sized appropriately for the target sequence length."""
-        if self.tokenizer is None:
-            self._load_tokenizer(dtype_override=dtype_override)
-
-        if seq_len not in PREFILL_TEXTS:
-            available = sorted(PREFILL_TEXTS.keys())
-            raise ValueError(
-                f"seq_len={seq_len} is not supported. Available sequence lengths: {available}"
-            )
-        texts = get_prefill_texts_for_batch(seq_len, batch_size)
-
-        inputs = self.tokenizer(
-            texts,
-            return_tensors="pt",
-            padding="max_length",
-            truncation=True,
-            max_length=seq_len,
-        )
-
-        if dtype_override is not None:
-            for key in inputs:
-                inputs[key] = cast_input_to_type(inputs[key], dtype_override)
-
-        self.seq_len = seq_len
-        return inputs
 
     def load_inputs_decode(self, dtype_override=None, batch_size=1):
         """Load decode-step inputs (single token + static KV cache).
